Drop EoL Debian, adjust deps for Puppet 6 & 7
Module dependencies were updated minimum versions that support Puppet 6
and max versions that includes the latest releases.

The "description" key was removed as it is no longer supported.

yumrepo_core was also dropped as it is only needed in the fixtures file.

yum was dropped because it is not used within this module